Fix incorrect import of assume
The `assume` function is [incorrectly][1] being imported from `hypothesis.strategies` instead of `hypothesis`.

This worked by coincidence in the [version][2] of hypothesis available when hypothesis-geojson was released.  In modern releases of hypothesis, assume is no longer imported into strategies, causing an import error which prevents usage of this library.  Since assume was not listed as a member of __all__, the import was inappropriate at the time of release anyway.

Correcting the import to the documented location fixes the import error.

[1]: https://hypothesis.readthedocs.io/en/latest/details.html#hypothesis.assume
[2]: https://github.com/HypothesisWorks/hypothesis/blob/3.11.6/src/hypothesis/strategies.py#L27